refactor/split-contact-component
